// src/app/[locale]/manage/orders/dishes-dialog.tsx
import { Button } from '@/components/ui/button'
import { Dialog, DialogContent, DialogHeader, DialogTitle, DialogTrigger } from '@/components/ui/dialog'
import { Table, TableBody, TableCell, TableHead, TableHeader, TableRow } from '@/components/ui/table'
import AutoPagination from '@/components/auto-pagination'
import { DishListResType } from '@/schemaValidations/dish.schema'
import { useEffect, useState } from 'react'
import {
  ColumnDef,
  ColumnFiltersState,
  SortingState,
  VisibilityState,
  flexRender,
  getCoreRowModel,
  getFilteredRowModel,
  getPaginationRowModel,
  getSortedRowModel,
  useReactTable
} from '@tanstack/react-table'
import { formatCurrency, getVietnameseDishStatus, simpleMatchText } from '@/lib/utils'
import { Input } from '@/components/ui/input'
import Image from 'next/image'
import { useGetDishList } from '@/queries/useDish'
import { useTranslations } from 'next-intl'

type DishItem = DishListResType['data'][0]

const PAGE_SIZE = 10
export function DishesDialog({ onChoose }: { onChoose: (dish: DishItem) => void }) {
  const t = useTranslations('ManageOrders')
  const tDishStatus = useTranslations('DishStatusLabel')
  const [open, setOpen] = useState(false)
  const dishListQuery = useGetDishList()
  const data = dishListQuery.data?.payload.data || []
  const columns: ColumnDef<DishItem>[] = [
    {
      id: 'dishName',
      header: t('columnDish'),
      cell: ({ row }) => (
        <div className='flex items-center space-x-4'>
          <Image
            src={row.original.image}
            alt={row.original.name}
            width={50}
            height={50}
            className='rounded-md object-cover w-[50px] h-[50px]'
          />
          <span>{row.original.name}</span>
        </div>
      ),
      filterFn: (row, columnId, filterValue: string) => {
        if (filterValue === undefined) return true
        return simpleMatchText(String(row.original.name), String(filterValue))
      }
    },
    {
      accessorKey: 'price',
      header: t('columnPrice'),
      cell: ({ row }) => <div className='capitalize'>{formatCurrency(row.getValue('price'))}</div>
    },
    {
      accessorKey: 'status',
      header: t('columnStatus'),
      cell: ({ row }) => <div>{tDishStatus(row.getValue<string>('status'))}</div>
    }
  ]
  const [sorting, setSorting] = useState<SortingState>([])
  const [columnFilters, setColumnFilters] = useState<ColumnFiltersState>([])
  const [columnVisibility, setColumnVisibility] = useState<VisibilityState>({})
  const [rowSelection, setRowSelection] = useState({})
  const [pagination, setPagination] = useState({
    pageIndex: 0, // Gía trị mặc định ban đầu, không có ý nghĩa khi data được fetch bất đồng bộ
    pageSize: PAGE_SIZE //default page size
  })

  const table = useReactTable({
    data,
    columns,
    onSortingChange: setSorting,
    onColumnFiltersChange: setColumnFilters,
    getCoreRowModel: getCoreRowModel(),
    getPaginationRowModel: getPaginationRowModel(),
    getSortedRowModel: getSortedRowModel(),
    getFilteredRowModel: getFilteredRowModel(),
    onColumnVisibilityChange: setColumnVisibility,
    onRowSelectionChange: setRowSelection,
    onPaginationChange: setPagination,
    autoResetPageIndex: false,
    state: {
      sorting,
      columnFilters,
      columnVisibility,
      rowSelection,
      pagination
    }
  })

  useEffect(() => {
    table.setPagination({
      pageIndex: 0,
      pageSize: PAGE_SIZE
    })
  }, [table])

  const choose = (dish: DishItem) => {
    onChoose(dish)
    setOpen(false)
  }

  return (
    <Dialog open={open} onOpenChange={setOpen}>
      <DialogTrigger asChild>
        <Button variant='outline'>{t('changeDish')}</Button>
      </DialogTrigger>
      <DialogContent className='sm:max-w-[600px] max-h-full overflow-auto'>
        <DialogHeader>
          <DialogTitle>{t('selectDish')}</DialogTitle>
        </DialogHeader>
        <div>
          <div className='w-full'>
            <div className='flex items-center py-4'>
              <Input
                placeholder={t('filterDishName')}
                value={(table.getColumn('dishName')?.getFilterValue() as string) ?? ''}
                onChange={(event) => table.getColumn('dishName')?.setFilterValue(event.target.value)}
                className='max-w-sm'
              />
            </div>
            <div className='rounded-md border'>
              <Table>
                <TableHeader>
                  {table.getHeaderGroups().map((headerGroup) => (
                    <TableRow key={headerGroup.id}>
                      {headerGroup.headers.map((header) => {
                        return (
                          <TableHead key={header.id}>
                            {header.isPlaceholder
                              ? null
                              : flexRender(header.column.columnDef.header, header.getContext())}
                          </TableHead>
                        )
                      })}
                    </TableRow>
                  ))}
                </TableHeader>
                <TableBody>
                  {table.getRowModel().rows?.length ? (
                    table.getRowModel().rows.map((row) => (
                      <TableRow
                        key={row.id}
                        data-state={row.getIsSelected() && 'selected'}
                        onClick={() => choose(row.original)}
                        className='cursor-pointer'
                      >
                        {row.getVisibleCells().map((cell) => (
                          <TableCell key={cell.id}>
                            {flexRender(cell.column.columnDef.cell, cell.getContext())}
                          </TableCell>
                        ))}
                      </TableRow>
                    ))
                  ) : (
                    <TableRow>
                      <TableCell colSpan={columns.length} className='h-24 text-center'>
                        No results.
                      </TableCell>
                    </TableRow>
                  )}
                </TableBody>
              </Table>
            </div>
            <div className='flex items-center justify-end space-x-2 py-4'>
              <div className='text-xs text-muted-foreground py-4 flex-1 '>
                {t('showing')} <strong>{table.getPaginationRowModel().rows.length}</strong> {t('of')}{' '}
                <strong>{data.length}</strong> {t('results')}
              </div>
              <div>
                <AutoPagination
                  page={table.getState().pagination.pageIndex + 1}
                  pageSize={table.getPageCount()}
                  onClick={(pageNumber) =>
                    table.setPagination({
                      pageIndex: pageNumber - 1,
                      pageSize: PAGE_SIZE,
                    })}
                  isLink={false}
                />
              </div>
            </div>
          </div>
        </div>
      </DialogContent>
    </Dialog>
  )
}
